Refactor folder creation and path handling
