+++ packages/nuxt/src/runtime/modules/drizzle/index.ts
@@ -7,7 +7,7 @@ import { basename } from 'pathe'
 import { definePergelModule } from '../../core/definePergel'
 import { useNitroImports } from '../../core/utils/useImports'
 import { globsBuilderWatch } from '../../core/utils/globs'
-import { createFolderModule } from '../../core/utils/useOpenFolder'
+import { createFolderModule } from '../../core/utils/createFolderModule'
 import type { DrizzleConfig, ResolvedDrizzleConfig } from './types'
 import { setupPostgres } from './drivers/postgres'
 import { copyMigrationFolder } from './core'
