+++ packages/nuxt/src/runtime/modules/ses/index.ts
@@ -13,10 +13,10 @@ export default definePergelModule({
   },
   defaults: {
   },
-  async setup({ nuxt, moduleOptions }) {
+  async setup({ nuxt, options }) {
     const resolver = createResolver(import.meta.url)
 
-    generateModuleRuntimeConfig<SesModuleRuntimeConfig>(nuxt, moduleOptions, {
+    generateModuleRuntimeConfig<SesModuleRuntimeConfig>(nuxt, options, {
       region: '',
       accessKeyId: '',
       secretAccessKey: '',
@@ -24,8 +24,8 @@ export default definePergelModule({
 
     addServerImportsDir(resolver.resolve('./composables'))
     nuxt._pergel.contents.push({
-      moduleName: moduleOptions.moduleName,
-      projectName: moduleOptions.projectName,
+      moduleName: options.moduleName,
+      projectName: options.projectName,
       content: /* ts */`
         function ses() {
           return {
